move epclass logic into config
