+++ src/energy.html
@@ -12,7 +12,6 @@
                 <script src="localization.js"></script>
                 <script src="energy.js"></script>
                 <script src="hourlyModel.js"></script>
-                <script src="epclass.js"></script>
                 <script src="glue.js" defer ></script>
         </head>
 	<body>
