+++ src/energyprint.html
@@ -396,7 +396,7 @@ margin-left: calc(8em + 0.5em);
     <!-- Print Button -->
     <button id="printButton" onclick="window.print()">Skriv ut</button>
 
-    <script src="epclass.js"></script>
+    <script src="config.js"></script>
     <script src="energyprint.js"></script>
   </body>
 </html>
