+++ src/energyprint_new.html
@@ -187,7 +187,6 @@
     </div>
   </div>
 
-  <script src="epclass.js"></script>
   <script src="config.js"></script>
   <script src="strings.js"></script>
   <script src="localization.js"></script>
